requirements.txt
functions-framework==3.*
aiohttp
aspose-words==23.10.0
beautifulsoup4
opencv-contrib-python-headless
numpy
openvino
Pillow
pillow-avif-plugin
PyMySQL==1.0.2
SQLAlchemy==1.4.9
scikit-image
scikit-learn
scipy
scrapfly-sdk
regex
requests
supabase
matplotlib
google-cloud-secret-manager
google-cloud-storage
w3lib
torch
textdistance
google-cloud-vision==3.7.2
tenacity
ultralytics 
